triggerCheck moved to command.go from router.go; test added
